# services/agentatlas/api/openapi/atlas-agent.yaml
openapi: 3.0.3
info:
  title: AgentAtlas Agent Control Plane API
  version: 1.0.0
  description: >
    Agent-first control plane for admins and knowledge owners: Knowledge Agent
    runs, workflow drafts and publication, dream policies, and artifact
    processing jobs. Drafts may be created by the agent; publication always
    requires an explicit human confirmation. Every call authenticates with an
    AgentNexus-issued ticket.
security:
  - NexusTicket: []
paths:
...
  /auth/login:
    get:
      security: []
      operationId: beginAtlasBrowserLogin
      parameters: [{name: return_to, in: query, schema: {type: string, pattern: '^/[^\\]*$'}}]
      responses: {'302': {description: Full-page redirect to AgentNexus OIDC authorization.}}
  /auth/callback:
    get:
      security: []
      operationId: completeAtlasBrowserLogin
      parameters:
        - {name: state, in: query, required: true, schema: {type: string}}
        - {name: code, in: query, required: true, schema: {type: string}}
      responses: {'302': {description: Opaque Atlas session created and original relative path restored.}}
  /api/session:
    get:
      security: [{AtlasSession: []}]
      operationId: getAtlasBrowserSession
      responses: {'200': {description: Browser identity and sealed scope snapshot, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserSessionView'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}}
  /api/knowledge:
    get:
      security: [{AtlasSession: []}]
      operationId: listBrowserKnowledgeHome
      description: Lists human-readable knowledge content after exact organization authorization by AgentNexus.
      parameters:
        - {name: org_unit_id, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
        - {name: query, in: query, schema: {type: string, maxLength: 200}}
      responses:
        '200': {description: Bounded knowledge workspace for one authorized organization, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserKnowledgeHome'}}}}
        '400': {description: The search query exceeds the supported length, content: {application/json: {schema: {$ref: '#/components/schemas/InvalidKnowledgeQueryError'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {description: The selected organization is not authorized.}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {description: Authorization or safely bounded knowledge data is unavailable.}
  /api/assessments/{id}:
    get:
      security: [{AtlasSession: []}]
      operationId: getManagerAssessmentDetail
      description: Returns the authorized manager assessment detail (score/level and manager notes) only for a subject within the manager's exact org hierarchy scope and under a policy they own. It exposes no HR-action capability.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, minLength: 1, maxLength: 256}}
      responses:
        '200': {description: The authorized manager assessment detail, content: {application/json: {schema: {type: object}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {description: The manager is outside the subject's reporting line or does not own the governing policy.}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {description: The assessment surface is not configured.}
  /api/dream/runs:
    get:
      security: [{AtlasSession: []}]
      operationId: listBrowserDreamRuns
      description: Lists immutable Dream runs after exact browser-session organization authorization by AgentNexus.
      parameters:
        - {name: org_unit_id, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
        - {name: window, in: query, schema: {type: string, format: date}}
      responses:
        '200': {description: Bounded immutable Dream runs, content: {application/json: {schema: {$ref: '#/components/schemas/DreamRunList'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {description: Browser session or AgentNexus denied the exact organization.}
        '503': {description: Authorization or bounded Dream data is unavailable.}
  /api/dream/runs/{id}:
    get:
      security: [{AtlasSession: []}]
      operationId: getBrowserDreamRun
      parameters: [{name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}]
      responses:
        '200': {description: Immutable sanitized Dream run, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamRunView'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {description: Browser session or AgentNexus denied the exact run organization.}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/dream/runs/{id}/annotations:
    post:
      security: [{AtlasSession: []}]
      operationId: annotateBrowserDreamRun
      description: Appends an idempotent governed annotation; immutable Dream output is never edited.
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 256}}
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [action], properties: {action: {type: string, enum: [confirm, reject, mark_incorrect, comment]}, comment: {type: string, maxLength: 4000}}}}}}
      responses: {'201': {description: Annotation appended, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamAnnotation'}}}}, '403': {description: Exact annotation authorization denied.}, '409': {description: Idempotency conflict.}}
  /api/dream/runs/{id}/reruns:
    post:
      security: [{AtlasSession: []}]
      operationId: rerunBrowserDreamRun
      description: Creates an idempotent immutable rerun pinned to the original run window and versions.
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 256}}
      responses: {'202': {description: Rerun scheduled, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamRunHandleResponse'}}}}, '403': {description: Exact rerun authorization denied.}, '409': {description: Idempotency conflict.}}
  /api/dream/runs/{id}/evidence-access:
    post:
      security: [{AtlasSession: []}]
      operationId: accessBrowserDreamEvidence
      description: Requests a resource-bound AgentNexus Step Grant, appends mandatory audit, then returns sanitized detail without raw pointer, grant, or audit identifiers.
      parameters: [{name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}]
      responses:
        '200': {description: Sanitized audited evidence detail, content: {application/json: {schema: {type: object, additionalProperties: false, required: [content_type, sanitized_detail, content_hash], properties: {content_type: {type: string}, sanitized_detail: {type: string, maxLength: 100000}, content_hash: {type: string}}}}}}
        '403': {description: Step Grant or exact organization authorization denied.}
        '503': {description: Authorization, Step Grant, evidence read, or mandatory audit unavailable.}
  /api/dream/policies:
    get:
      security: [{AtlasSession: []}]
      operationId: listBrowserDreamPolicies
      parameters: [{name: org_unit_id, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 128}}]
      responses: {'200': {description: Bounded sanitized lifecycle views, content: {application/json: {schema: {type: object, additionalProperties: false, required: [dream_policies], properties: {dream_policies: {type: array, maxItems: 1000, items: {$ref: '#/components/schemas/BrowserDreamPolicyLifecycle'}}}}}}}, '403': {description: Exact policy read authorization denied.}}
    post:
      security: [{AtlasSession: []}]
      operationId: createBrowserDreamPolicyDraft
      description: Creates an audited governed draft only; this route never publishes.
      parameters: [{name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 256}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamBasicPolicyInput'}}}}
      responses: {'201': {description: Governed Dream policy draft created, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamPolicyLifecycle'}}}}, '403': {description: Exact edit or suggestion authorization denied.}, '409': {description: Idempotency conflict.}, '422': {description: Basic policy input invalid.}}
  /api/dream/workflow-bindings:
    get:
      security: [{AtlasSession: []}]
      operationId: listBrowserDreamWorkflowBindings
      description: Lists human-readable published Dream workflows with opaque session-bound binding handles; workflow and output-space identifiers are never returned.
      parameters: [{name: org_unit_id, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 128}}]
      responses:
        '200': {description: Bounded published Dream workflow bindings, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamWorkflowBindingList'}}}}
        '403': {description: Exact organization authorization denied.}
        '503': {description: Published workflow or authoritative output binding unavailable.}
  /api/dream/policies/{id}:
    put:
      security: [{AtlasSession: []}]
      operationId: updateBrowserDreamPolicyDraft
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamBasicPolicyUpdate'}}}}
      responses: {'200': {description: Draft updated without publication, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamPolicyLifecycle'}}}}, '403': {description: Exact edit authorization denied.}, '409': {description: Revision or idempotency conflict.}}
  /api/dream/policies/{id}/advanced:
    get:
      security: [{AtlasSession: []}]
      operationId: getBrowserDreamAdvancedPolicy
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: X-Atlas-Advanced-Mode, in: header, required: true, schema: {type: string, enum: [enabled]}}
      responses: {'200': {description: Advanced policy fields, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamAdvancedPolicy'}}}}, '403': {description: Advanced entitlement, mode, or exact authorization denied.}}
    put:
      security: [{AtlasSession: []}]
      operationId: updateBrowserDreamAdvancedPolicy
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: X-Atlas-Advanced-Mode, in: header, required: true, schema: {type: string, enum: [enabled]}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamAdvancedPolicy'}}}}
      responses: {'200': {description: Advanced policy draft updated, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamPolicyLifecycle'}}}}, '403': {description: Advanced entitlement, mode, or exact authorization denied.}, '409': {description: Revision or idempotency conflict.}, '422': {description: Invalid advanced policy.}}
  /api/dream/policies/{id}/check:
    post:
      security: [{AtlasSession: []}]
      operationId: checkBrowserDreamPolicyRisk
      parameters: [{name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [revision], properties: {revision: {type: integer, minimum: 0}}}}}}
      responses: {'200': {description: Deterministic risk and changed fields.}, '403': {description: Exact edit authorization denied.}, '409': {description: Stale revision.}}
  /api/dream/policies/{id}/adoptions:
    post:
      security: [{AtlasSession: []}]
      operationId: adoptBrowserDreamPolicySuggestion
      description: Creates a new direct-edit draft from an immutable employee suggestion and records durable source lineage; the source suggestion is never mutated.
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [revision], properties: {revision: {type: integer, minimum: 0}}}}}}
      responses:
        '201': {description: New direct-edit draft created with immutable adoption lineage, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamPolicyLifecycle'}}}}
        '403': {description: Exact edit or organization authorization denied.}
        '409': {description: Suggestion state, revision, concurrency, or idempotency conflict.}
  /api/dream/policies/{id}/review:
    post:
      security: [{AtlasSession: []}]
      operationId: submitBrowserDreamPolicyReview
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [revision, action], properties: {revision: {type: integer, minimum: 0}, action: {type: string, enum: [publish, disable]}}}}}}
      responses: {'200': {description: AgentNexus-bound single-confirmation, upward-review, or admin-queue route recorded.}, '403': {description: Requester, organization, or self-review rule denied.}, '409': {description: Invalid lifecycle state or revision.}}
  /api/dream/policies/{id}/decisions:
    post:
      security: [{AtlasSession: []}]
      operationId: decideBrowserDreamPolicy
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [revision, decision], properties: {revision: {type: integer, minimum: 0}, decision: {type: string, enum: [approve, reject]}, comment: {type: string, maxLength: 4000}}}}}}
      responses: {'200': {description: Assigned decision recorded; publication still separate.}, '403': {description: Wrong reviewer, self-review, permission, or organization denied.}, '409': {description: Invalid lifecycle state or revision.}}
  /api/dream/policies/{id}/publish:
    post:
      security: [{AtlasSession: []}]
      operationId: publishBrowserDreamPolicy
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [revision], properties: {revision: {type: integer, minimum: 0}}}}}}
      responses: {'200': {description: Approved policy version published atomically.}, '403': {description: Exact edit authorization denied.}, '409': {description: Approved decision, valid workflow, revision, or idempotency requirement failed.}}
  /api/dream/policies/{id}/disable:
    post:
      security: [{AtlasSession: []}]
      operationId: disableBrowserDreamPolicy
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [revision], properties: {revision: {type: integer, minimum: 0}}}}}}
      responses: {'200': {description: Approved high-risk disable completed.}, '403': {description: Exact edit authorization denied.}, '409': {description: Approved disable decision or lifecycle requirement failed.}}
  /api/dream/policies/{id}/backfills:
    post:
      security: [{AtlasSession: []}]
      operationId: backfillBrowserDreamPolicy
      parameters:
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueDreamHandle'}}
        - {name: X-Atlas-Advanced-Mode, in: header, required: true, schema: {type: string, enum: [enabled]}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamBackfillRequest'}}}}
      responses: {'202': {description: Bounded audited backfill accepted, content: {application/json: {schema: {$ref: '#/components/schemas/BrowserDreamRunHandleResponse'}}}}, '403': {description: Advanced maintenance or exact edit authorization denied.}, '409': {description: Window, lineage, overlap, or idempotency conflict.}}
  /auth/logout:
    post:
      security: [{AtlasSession: []}]
      operationId: logoutAtlasBrowserSession
      responses: {'204': {description: Atlas and AgentNexus sessions revoked.}}
  /api/legacy/knowledge:
    get:
      security: [{AtlasSession: []}]
      operationId: listLegacyKnowledge
      responses: {'200': {$ref: '#/components/responses/LegacyItems'}, '403': {description: Advanced maintenance or organization authorization denied.}, '503': {description: Safe compatibility adapter unavailable.}}
  /api/legacy/dream:
    get:
      security: [{AtlasSession: []}]
      operationId: listLegacyDream
      responses: {'200': {$ref: '#/components/responses/LegacyItems'}, '403': {description: Advanced maintenance or organization authorization denied.}, '503': {description: Safe compatibility adapter unavailable.}}
  /api/legacy/workflows:
    get:
      security: [{AtlasSession: []}]
      operationId: listLegacyWorkflows
      responses: {'200': {$ref: '#/components/responses/LegacyItems'}, '403': {description: Advanced maintenance or organization authorization denied.}, '503': {description: Safe compatibility adapter unavailable.}}
...
  /api/legacy/assistant/attachments:
    post:
      security: [{AtlasSession: []}]
      operationId: uploadLegacyAssistantAttachments
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files: {type: array, maxItems: 20, items: {type: string, format: binary}}
      responses: {'200': {description: Attachments accepted by the secure adapter.}, '403': {description: Advanced maintenance or organization authorization denied.}, '503': {description: Safe attachment adapter unavailable; no file is consumed.}}
  /api/changes/suggestions:
    post:
      security: [{AtlasSession: []}]
      operationId: suggestGovernedChange
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SuggestionInput'}}}}
      responses: {'201': {description: Suggestion draft, content: {application/json: {schema: {$ref: '#/components/schemas/ChangeDraft'}}}}, '403': {description: Suggestion denied.}}
...
components:
  securitySchemes:
    NexusTicket:
      type: apiKey
      in: header
      name: X-Nexus-Ticket
    AtlasSession:
      type: apiKey
      in: cookie
      name: atlas_session
...
  schemas:
...
    BrowserSessionView:
      type: object
      additionalProperties: false
      required: [authenticated, enterprise_id, enterprise_name, enterprise_user_id, display_name, org_version, org_unit_ids, org_tree, permissions]
      properties:
        authenticated: {type: boolean}
        enterprise_id: {type: string}
        enterprise_name: {type: string, minLength: 1}
        enterprise_user_id: {type: string}
        display_name: {type: string}
        org_version: {type: integer, format: int64}
        org_unit_ids: {type: array, maxItems: 1000, items: {type: string}}
        org_tree: {type: array, maxItems: 1000, items: {$ref: '#/components/schemas/BrowserOrgScopeNode'}}
        permissions: {type: array, maxItems: 100, items: {type: string}}
        advanced_mode_allowed: {type: boolean}
        idle_expires_at: {type: string, format: date-time}
        absolute_expires_at: {type: string, format: date-time}
...